Throw error on invalid date
