+++ src/register-commands.ts
@@ -4,8 +4,21 @@ import { ApiProvider } from "./api-provider";
 import { start } from "./commands/start";
 import { finish } from "./commands/finish";
 import { auth } from "./commands/auth";
-import { parseDateOption } from "./utils/parse-date-option";
+import { isValidDate } from "./utils/is-valid-date";
 import { getTodayDate } from "./utils/get-today-date";
+import { log } from "./utils/log";
+
+const getDateForCommand = (command) => {
+    if (command.date === undefined || command.date === null) {
+        return getTodayDate();
+    }
+
+    if (isValidDate(command.date)) {
+        return command.date;
+    }
+
+    return null;
+};
 
 export const registerCommands = (commander: CommanderStatic, apiProvider: ApiProvider) => {
     commander
@@ -13,7 +26,12 @@ export const registerCommands = (commander: CommanderStatic, apiProvider: ApiPro
         .description("start a timer")
         .option("-d, --date <date>", "specify a date in YYYY-MM-DD format")
         .action((cmd) => {
-            const date = parseDateOption(getTodayDate, cmd.date);
+            const date = getDateForCommand(cmd);
+
+            if (date === null) {
+                log.error("Invalid date provided. Use YYYY-MM-DD");
+                process.exit(1);
+            }
 
             return start(apiProvider, date);
         });
@@ -23,7 +41,12 @@ export const registerCommands = (commander: CommanderStatic, apiProvider: ApiPro
         .description("finish a timer")
         .option("-d, --date <date>", "specify a date in YYYY-MM-DD format")
         .action((cmd) => {
-            const date = parseDateOption(getTodayDate, cmd.date);
+            const date = getDateForCommand(cmd);
+
+            if (date === null) {
+                log.error("Invalid date provided. Use YYYY-MM-DD");
+                process.exit(1);
+            }
 
             return finish(apiProvider, date);
         });
