+++ src/utils/is-valid-date.ts
@@ -1,0 +1,3 @@
+import * as moment from "moment";
+
+export const isValidDate = (date: string) => moment(date, "YYYY-MM-DD", true).isValid();
